refactor PlasmaCliParamsParserModule
