+++ cli/src/main/scala/org/plasmalabs/cli/params/package.scala
@@ -1,0 +1,7 @@
+package org.plasmalabs.cli
+
+package object params:
+
+  object implicits extends ReadInstances
+
+end params
